assessment/src/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.0
Markdown==3.7
psycopg==3.2.3
python-environ==0.4.54
sqlparse==0.5.1
typing_extensions==4.12.2
uWSGI==2.0.28
